backend/requirements.txt
# FastAPI + server
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20

# Database
supabase==2.15.1

# AWS Bedrock
boto3==1.38.19

# Google Vertex AI — new unified SDK
google-genai==1.10.0

# Data processing
pandas==2.2.3

# Utilities
python-dotenv==1.1.0
httpx==0.28.1
